requirements.txt
# Core Dependencies
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0

# API Integrations  
praw>=7.7.0
newsapi-python>=0.2.6
requests>=2.31.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Utilities
python-dotenv>=1.0.0
